add: Homeを共通化するためにHomeRouteを追加した
