+++ src/resources/ts/App.tsx
@@ -5,7 +5,8 @@ import { Chats } from "@/ts/Chat/Chats";
 import { QueryClient, QueryClientProvider } from "react-query";
 import { Route } from "@/ts/layout/Route";
 import { createTheme, ThemeProvider } from "@mui/material";
-import { Home } from "@/ts/Home/Home";
+import {HomeRoute} from "@/ts/layout/HomeRoute";
+import {Chat} from "@/ts/Home/Chat";
 
 export const App = () => {
     const queryClient = new QueryClient();
@@ -16,8 +17,8 @@ export const App = () => {
             <ThemeProvider theme={theme}>
                 <BrowserRouter>
                     <Route path="login" element={<Login/>}/>
-                    <Route path="/" element={<Home/>}/>
-                    <Route path="chats" element={<Chats/>}/>
+                    <HomeRoute path="/" element={<Chat/>}/>
+                    <HomeRoute path="chats" element={<Chats/>}/>
                 </BrowserRouter>
             </ThemeProvider>
         </QueryClientProvider>
