+++ src/resources/ts/layout/Header.tsx
@@ -1,5 +1,6 @@
 import {
     AppBar,
+    Box,
     Button,
     Container,
     Toolbar,
@@ -32,3 +33,5 @@ export const Header = () => {
         </AppBar>
     )
 };
+
+export const Spacer = () => <Box sx={(theme) => ({...theme.mixins.toolbar})} />;
