+++ src/resources/ts/layout/Route.tsx
@@ -4,38 +4,27 @@ import {
     RouteProps,
 } from "react-router-dom";
 import React from "react";
-import { Header } from "@/ts/layout/Header";
-import { Box, Container, CssBaseline } from "@mui/material";
-import { useAppBarHeight } from "@/ts/layout/useAppBarHeight";
+import {Header, Spacer} from "@/ts/layout/Header";
+import { Container, CssBaseline } from "@mui/material";
 
-export const Route = (props: RouteProps) => {
-    // https://github.com/mui/material-ui/issues/10739
-    const appBarMinHeight = useAppBarHeight();
-
-    return (
-        <Routes>
-            <BaseRoute
-                {...props}
-                element={
-                    <>
-                        <CssBaseline/>
-                        <Header/>
-                        <Spacer/>
-                        <Container
-                            component="main"
-                            maxWidth={false}
-                            disableGutters
-                            sx={{
-                                height: `calc(100vh - ${appBarMinHeight}px)`,
-                            }}
-                        >
-                            {props.element}
-                        </Container>
-                    </>
-                }
-            />
-        </Routes>
-    );
-};
-
-const Spacer = () => <Box sx={(theme) => ({...theme.mixins.toolbar})} />;
+export const Route = (props: RouteProps) => (
+    <Routes>
+        <BaseRoute
+            {...props}
+            element={
+                <>
+                    <CssBaseline/>
+                    <Header/>
+                    <Spacer/>
+                    <Container
+                        component="main"
+                        maxWidth={false}
+                        disableGutters
+                    >
+                        {props.element}
+                    </Container>
+                </>
+            }
+        />
+    </Routes>
+);
